Pulizia del codice

+++ src/main/java/com/services/CeramicheService.java
@@ -38,7 +38,7 @@ public class CeramicheService {
 
     //SEZIONE DELETE
     public void deleteCeramicaItemById(Integer id) throws NotFoundException {
-        if(repository.existsById(id)){
+        if(repository.existsById(id) && repository.findById(id).isPresent()){
             repository.delete(repository.findById(id).get());
         }
         throw new NotFoundException("Item not found");
